print function

+++ print_function.c
@@ -101,7 +101,7 @@ int print_i(va_list args)
 
 int print_b(va_list args)
 {
-	long int number = va_arg(args, int), b = 1, lengh = 0, number1;
+	int number = va_arg(args, int), b = 1, lengh = 0, number1;
 
 	number1 = convert(number, 2);
 	if (number1 < 0)
